--- src/utils/compose.js
@@ -1,7 +1,0 @@
-export const compose = (filters, ...utilFunctions) => {
-  return (initialProducts) =>
-    utilFunctions.reduce(
-      (filteredProducts, func) => func(filteredProducts,filters),
-      initialProducts
-    );
-};
