--- src/utils/getDiscountPercentage.js
@@ -1,2 +1,0 @@
-export const getDiscountPercentage = (price, offer_price) =>
-  Math.round(((price - offer_price) / price) * 100);
