--- src/utils/productSearchFilter.js
@@ -1,6 +1,0 @@
-export const productSearchFilter = (products, filters) => {
-  if (!filters.searchQuery) return products;
-  return products.filter((product) =>
-    product.title.toLowerCase().includes(filters.searchQuery.toLowerCase())
-  );
-};
