--- src/utils/ratingFilter.js
@@ -1,6 +1,0 @@
-export const ratingFilter = (products, filters) => {
-  if(filters.ratingFilterValue === null) return products;
-  return products.filter((product) => {
-    return Number(product.rating) >= filters.ratingFilterValue;
-  });
-};
